Move dictionaries needed for testing to new file
This ensures that we can get at those constants without having to have
test dependencies (e.g. pytest) installed.
